requirements.txt
aiogram~=3.3.0
python-dotenv~=1.0.1
envparse~=0.2.0
loguru~=0.7.2
tortoise-orm~=0.20.0
asyncpg~=0.29.0
Babel~=2.14.0
pygsheets~=2.0.6